Fixed a Few Style Things that were Discussed This commit: ~~Also did a few undiscussed but need approval Because: --{Stopping for the Evening}
