--- src/components/SeniorCard.js
@@ -1,1 +1,0 @@
-import React from 'react';
