+++ src/components/senior/SeniorDash.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styled from 'styled-components'
+import styled from 'styled-components';
 
 const BackgroundDiv = styled.div`
 	height: 100vh;
@@ -11,13 +11,19 @@ const BackgroundDiv = styled.div`
 `;
 
 const SeniorDash = () => {
-  return (
-    <BackgroundDiv>
-      <div><h1>To Do</h1></div>
-      <div><p>Body of Todo List</p></div>
-      <div><span>Extra Space</span></div>
-    </BackgroundDiv>
-  )
-}
+	return (
+		<BackgroundDiv>
+			<div>
+				<h1>To Do</h1>
+			</div>
+			<div>
+				<p>Body of Todo List</p>
+			</div>
+			<div>
+				<span>Extra Space</span>
+			</div>
+		</BackgroundDiv>
+	);
+};
 
-export default SeniorDash
+export default SeniorDash;
