+++ src/components/senior/SeniorList.js
@@ -1,0 +1,1 @@
+import React from 'react';
