+++ src/components/student/StudentSignUp.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import { postNewStudent } from '../redux/actions';
+import { postNewStudent } from '../../redux/actions';
 import { Button, Container, Row, Col } from 'reactstrap';
 import { AvForm, AvField } from 'availity-reactstrap-validation';
 import styled from 'styled-components';
@@ -47,14 +47,13 @@ const StudentSignUp = () => {
 		history.push('/student-dash');
 	};
 
-		
 	return (
 		<BackgroundDiv>
 			<Container style={{ paddingTop: '40px' }}>
 				<Row>
 					<Col sm='12' md={{ size: 6, offset: 3 }}>
-						<h1>Student Sign Up</h1>
-						<AvForm className='StudentSignUp-form' onSubmit={handleSubmit} style={{ marginTop: '40px' }}>
+						<h1 className='nonForm'>Student Sign Up</h1>
+						<AvForm className='formWrapper' onSubmit={handleSubmit} style={{ marginTop: '40px' }}>
 							<AvField
 								label='First Name'
 								type='text'
@@ -111,7 +110,9 @@ const StudentSignUp = () => {
 									},
 								}}
 							/>
-							<Button type='submit'>Sign Up</Button>
+							<Button className='formButton' type='submit'>
+								Sign Up
+							</Button>
 						</AvForm>
 					</Col>
 				</Row>
